+++ source/extensions/omni.isaac.ros2_bridge/python/tests/__init__.py
@@ -12,6 +12,7 @@ import sys
 from .test_camera import *
 from .test_clock import *
 from .test_differential_base import *
+from .test_joint_state import *
 from .test_lidar import *
 from .test_point_cloud import *
 from .test_pose_tree import *