+++ source/extensions/omni.isaac.ros2_bridge/rclpy/__init__.py
@@ -37,7 +37,14 @@ class Extension(omni.ext.IExt):
                 omni.kit.app.get_app().print_and_log(f"Could not import system rclpy: {e}")
                 omni.kit.app.get_app().print_and_log("Attempting to load internal rclpy")
                 sys.path.append(os.path.join(os.path.dirname(__file__)))
-
+                if sys.platform == "win32":
+                    ext_manager = omni.kit.app.get_app().get_extension_manager()
+                    self._extension_path = ext_manager.get_extension_path(ext_id)
+                    if os.environ.get("PATH"):
+                        os.environ["PATH"] = os.environ.get("PATH") + ";" + self._extension_path + f"/{ros_distro}/lib"
+                    else:
+                        os.environ["PATH"] = self._extension_path + f"/{ros_distro}/lib"
+                        os.environ["RMW_IMPLEMENTATION"] = "rmw_fastrtps_cpp"
                 try:
                     import rclpy
 
@@ -46,8 +53,6 @@ class Extension(omni.ext.IExt):
                     omni.kit.app.get_app().print_and_log("rclpy loaded")
                 except Exception as e:
                     omni.kit.app.get_app().print_and_log(f"Could not import internal rclpy: {e}")
-                    ext_manager = omni.kit.app.get_app().get_extension_manager()
-                    self._extension_path = ext_manager.get_extension_path(ext_id)
                     if sys.platform == "linux":
                         omni.kit.app.get_app().print_and_log(
                             f"To use the Internal rclpy included with the extension please set: \nRMW_IMPLEMENTATION=rmw_fastrtps_cpp\nLD_LIBRARY_PATH=$LD_LIBRARY_PATH:{self._extension_path}/{ros_distro}/lib\nBefore starting Isaac Sim"