+++ source/extensions/omni.isaac.tests/omni/isaac/tests/tests/test_carter_v2.py
@@ -54,14 +54,8 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
 
         self._extension_path = get_extension_path_from_name("omni.isaac.tests")
 
-        ## setup carter_v2:
-        # open local carter_v2
-        # (result, error) = await omni.usd.get_context().open_stage_async(
-        #     self._extension_path + "/data/tests/carter_v2.usd"
-        # )
-
         # add in carter (from nucleus)
-        self.usd_path = self._assets_root_path + "/Isaac/Robots/Carter/carter_v2.usd"
+        self.usd_path = self._assets_root_path + "/Isaac/Robots/Carter/carter_v2_4.usd"
         (result, error) = await open_stage_async(self.usd_path)
 
         # Make sure the stage loaded
@@ -74,7 +68,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         # setup omnigraph
         self.graph_path = "/ActionGraph"
         graph, self.odom_node = setup_robot_og(
-            self.graph_path, "joint_wheel_left", "joint_wheel_right", "/carter_v2", 0.14, 0.4132
+            self.graph_path, "joint_wheel_left", "joint_wheel_right", "/World/Carter_V24", 0.14, 0.4132
         )
 
         pass
@@ -98,7 +92,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
 
         # get the dofbot
-        self.ar = self.dc.get_articulation("/carter_v2")
+        self.ar = self.dc.get_articulation("/World/Carter_V24")
         self.chassis = self.dc.get_articulation_root_body(self.ar)
         self.starting_pos = np.array(self.dc.get_rigid_body_pose(self.chassis).p)
 
@@ -129,7 +123,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/carter_v2")
+        await init_robot_sim(self.dc, "/World/Carter_V24")
 
         for x in range(1, 5):
             forward_velocity = x * 0.15
@@ -160,7 +154,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/carter_v2")
+        await init_robot_sim(self.dc, "/World/Carter_V24")
         for x in range(1, 5):
             self.my_world.play()
             await omni.kit.app.get_app().next_update_async()
@@ -190,7 +184,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         # Start Simulation and wait
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
-        await init_robot_sim(self.dc, "/carter_v2")
+        await init_robot_sim(self.dc, "/World/Carter_V24")
 
         for x in range(1, 4):
             angular_velocity = 0.6 * x
@@ -219,7 +213,7 @@ class TestCarterv2(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/carter_v2")
+        await init_robot_sim(self.dc, "/World/Carter_V24")
         forward_velocity = -0.1
         angular_velocity = -0.5
         og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(forward_velocity)