+++ source/standalone_examples/api/omni.isaac.jetbot/stable_baselines_example/eval.py
@@ -26,11 +26,11 @@ my_env = JetBotEnv(headless=False)
 model = PPO.load(policy_path)
 
 for _ in range(20):
-    obs = my_env.reset()
+    obs, _ = my_env.reset()
     done = False
     while not done:
         actions, _ = model.predict(observation=obs, deterministic=True)
-        obs, reward, done, info = my_env.step(actions)
+        obs, reward, done, truncated, info = my_env.step(actions)
         my_env.render()
 
 my_env.close()