+++ source/standalone_examples/api/omni.isaac.sensor/imu_sensor.py
@@ -16,10 +16,12 @@ import sys
 import carb
 import numpy as np
 from omni.isaac.core import World
+from omni.isaac.core.articulations import Articulation
 from omni.isaac.core.utils.nucleus import get_assets_root_path
+from omni.isaac.core.utils.stage import add_reference_to_stage
+from omni.isaac.core.utils.types import ArticulationAction
 from omni.isaac.sensor import IMUSensor
 from omni.isaac.wheeled_robots.controllers.differential_controller import DifferentialController
-from omni.isaac.wheeled_robots.robots import WheeledRobot
 
 my_world = World(stage_units_in_meters=1.0)
 my_world.scene.add_default_ground_plane()
@@ -29,23 +31,21 @@ if assets_root_path is None:
     carb.log_error("Could not find Isaac Sim assets folder")
     simulation_app.close()
     sys.exit()
-asset_path = assets_root_path + "/Isaac/Robots/Carter/carter_v2.usd"
+
+asset_path = assets_root_path + "/Isaac/Robots/Carter/carter_v2_4_sensors.usd"
+add_reference_to_stage(usd_path=asset_path, prim_path="/World/Carter")
+
 my_carter = my_world.scene.add(
-    WheeledRobot(
-        prim_path="/World/Carter",
-        name="my_carter",
-        wheel_dof_names=["joint_wheel_left", "joint_wheel_right"],
-        create_robot=True,
-        usd_path=asset_path,
-        position=np.array([0, 0.0, 0.5]),
-    )
+    Articulation(prim_path="/World/Carter/Carter_V24", name="my_carter", position=np.array([0, 0.0, 0.5]))
 )
+wheel_dof_names = ["joint_wheel_left", "joint_wheel_right"]
+
 my_controller = DifferentialController(name="simple_control", wheel_radius=0.04295, wheel_base=0.4132)
 
 
 imu_sensor = my_world.scene.add(
     IMUSensor(
-        prim_path="/World/Carter/caster_wheel_left/imu_sensor",
+        prim_path="/World/Carter/Carter_V24/caster_wheel_left/imu_sensor",
         name="imu",
         frequency=60,
         translation=np.array([0, 0, 0]),
@@ -56,20 +56,31 @@ i = 0
 while simulation_app.is_running():
     my_world.step(render=True)
     if my_world.is_playing():
+        wheel_dof_indices = [my_carter.get_dof_index(wheel_dof_names[i]) for i in range(len(wheel_dof_names))]
         if my_world.current_time_step_index == 0:
             my_world.reset()
             my_controller.reset()
         print(imu_sensor.get_current_frame())
+        actions = ArticulationAction()
         if i >= 0 and i < 1000:
             # forward
-            my_carter.apply_wheel_actions(my_controller.forward(command=[0.05, 0]))
+            actions = my_controller.forward(command=[0.05, 0])
+
         elif i >= 1000 and i < 1265:
             # rotate
-            my_carter.apply_wheel_actions(my_controller.forward(command=[0.0, np.pi / 12]))
+            actions = my_controller.forward(command=[0.0, np.pi / 12])
         elif i >= 1265 and i < 2000:
             # forward
-            my_carter.apply_wheel_actions(my_controller.forward(command=[0.05, 0]))
+            actions = my_controller.forward(command=[0.05, 0])
         elif i == 2000:
             i = 0
         i += 1
+        joint_actions = ArticulationAction()
+        joint_actions.joint_velocities = np.zeros(my_carter.num_dof)
+        if actions.joint_velocities is not None:
+            for j in range(len(wheel_dof_indices)):
+                joint_actions.joint_velocities[wheel_dof_indices[j]] = actions.joint_velocities[j]
+
+        my_carter.apply_action(joint_actions)
+
 simulation_app.close()