+++ source/extensions/omni.isaac.wheeled_robots/docs/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [0.7.1] - 2023-10-06
+### Changed
+- wheeled robot class can now accept a relative path from the default prim to the robot prim when the robot is not the default prim
+### Fixed
+- Differential controller now resets on simulation Stop
+
 ## [0.7.0] - 2023-09-06
 ### Added
 - Ackermann Steering OmniGraph node